+++ CHANGELOG.md
@@ -1,5 +1,14 @@
 # Changelog
 
+## dev
+
+### Fixed
+
+ * broken pipes (such as from `igseq something | something else`) are now
+   handled gracefully ([#30])
+
+[#30]: https://github.com/ShawHahnLab/igseq/pull/30
+
 ## 0.3.0 - 2022-07-14
 
 ### Added