+++ CHANGELOG.md
@@ -5,7 +5,7 @@
 ### Added
 
  * `identity` command for calculating pairwise identity between arbitrary
-   queries and references ([#31])
+   queries and references ([#31], [#37])
  * Support for showing basic tree topology for Newick-format files in `show`
    command ([#33])
 
@@ -15,6 +15,7 @@
  * broken pipes (such as from `igseq something | something else`) are now
    handled gracefully ([#30])
 
+[#37]: https://github.com/ShawHahnLab/igseq/pull/37
 [#35]: https://github.com/ShawHahnLab/igseq/pull/35
 [#33]: https://github.com/ShawHahnLab/igseq/pull/33
 [#31]: https://github.com/ShawHahnLab/igseq/pull/31